+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Hardcoded basepath removed
 - Removed 'Distributed by:' from bookdetails
 - Build and test:ci runs now contain linting
+- Transifex translations i18n library